+++ src/main/java/br/gov/es/participe/controller/HighlightController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -14,6 +15,7 @@ import br.gov.es.participe.controller.dto.HighlightDto;
 import br.gov.es.participe.controller.dto.HighlightParamDto;
 import br.gov.es.participe.model.Highlight;
 import br.gov.es.participe.service.HighlightService;
+import br.gov.es.participe.service.PersonService;
 
 @RestController
 @CrossOrigin
@@ -21,11 +23,21 @@ import br.gov.es.participe.service.HighlightService;
 public class HighlightController {
 
 	@Autowired
+	private PersonService personService;
+
+	@Autowired
 	private HighlightService highlightService;
 
 	@SuppressWarnings("rawtypes")
 	@PostMapping
-	public ResponseEntity store(@RequestBody HighlightParamDto highlightParamDto) {
+	public ResponseEntity store(
+			@RequestHeader("Authorization") String token,
+			@RequestBody HighlightParamDto highlightParamDto) {
+
+		if (!personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+			return ResponseEntity.status(401).body(null);
+		}
+
 		Highlight highlight = new Highlight(highlightParamDto);
 
 		HighlightDto response = new HighlightDto(highlightService.save(highlight, "rem"));
@@ -34,14 +46,24 @@ public class HighlightController {
 
 	@SuppressWarnings("rawtypes")
 	@DeleteMapping("/deleteAll/{id}")
-	public ResponseEntity deleteAll(@PathVariable Long id) {
+	public ResponseEntity deleteAll(
+			@RequestHeader("Authorization") String token,
+			@PathVariable Long id) {
+		if (!personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+			return ResponseEntity.status(401).body(null);
+		}
 		highlightService.deleteAllByIdPerson(id);
 		return ResponseEntity.status(200).build();
 	}
 
 	@SuppressWarnings("rawtypes")
 	@DeleteMapping
-	public ResponseEntity delete(@RequestBody HighlightParamDto highlightParamDto) {
+	public ResponseEntity delete(
+			@RequestHeader("Authorization") String token,
+			@RequestBody HighlightParamDto highlightParamDto) {
+		if (!personService.hasOneOfTheRoles(token, new String[] { "Administrator" })) {
+			return ResponseEntity.status(401).body(null);
+		}
 		boolean response = highlightService.delete(new Highlight(highlightParamDto));
 
 		return ResponseEntity.status(200).body(response);
